Fix 'NonHierarchicalClustering' object has no attribute 'ranges' ch5
In chapter 5 method compute_distance_matrix_instances the attribute self.ranges
is accessed but this is unset. The cause is that ranges above is set as variable
within the scope of the current method but since this is not the constructor of
the class it does not actually set self.ranges. This in turn causes the
subsequent access to result in an AttributeError.